feat: add modules solution
